+++ lua/telescope/_extensions/gh_actions.lua
@@ -1,0 +1,50 @@
+local actions = require('telescope.actions')
+local utils = require('telescope.utils')
+local Job = require('plenary.job')
+
+local A ={}
+-- a for actions
+A.gh_pr_checkout = function(prompt_bufnr)
+  local selection = actions.get_selected_entry(prompt_bufnr)
+  actions.close(prompt_bufnr)
+  local tmp_table = vim.split(selection.value,"\t");
+  if vim.tbl_isempty(tmp_table) then
+    return
+  end
+  local job = Job:new({
+      enable_recording = true ,
+      command = "gh",
+      args = {"pr", "checkout" ,tmp_table[1]}
+    })
+  -- need to display result in quickfix
+  job:sync()
+end
+
+
+A.gh_web_view=function(type)
+  return function(prompt_bufnr)
+    local selection = actions.get_selected_entry(prompt_bufnr)
+    actions.close(prompt_bufnr)
+    local tmp_table = vim.split(selection.value,"\t");
+    if vim.tbl_isempty(tmp_table) then
+      return
+    end
+    os.execute('gh ' .. type .. ' view --web ' .. tmp_table[1])
+  end
+end
+
+A.gh_gist_append=function(prompt_bufnr)
+  local selection = actions.get_selected_entry(prompt_bufnr)
+  actions.close(prompt_bufnr)
+  local tmp_table = vim.split(selection.value,"\t");
+  if vim.tbl_isempty(tmp_table) then
+    return
+  end
+  local gist_id = tmp_table[1]
+  local text = utils.get_os_command_output('gh gist view ' .. gist_id .. ' -r')
+  if text and vim.api.nvim_buf_get_option(vim.api.nvim_get_current_buf(), "modifiable") then
+    vim.api.nvim_put(vim.split(text,'\n'), 'b', true, true)
+  end
+end
+
+return A
